1036 - Fix updating user address
When the user is updated, the data received has an empty array for
the list of the user's community. The update order in the
`RestRepository` is such:

1. Save the User with filled fields
2. Save the relations from the received data
3. Save the User again.

With the code as it was we would update the address and communities in
1, sync the communities with the empty list of communities recieved in 2
(undoing part of what we did in 1),
and finally not re-update the community in 3 since the address is now
correct after being saved in 1.

With this fix (bringing the behavior back to what it was before !397),
we essentially update the location twice (in 1 and 3) which means doing
two geocoding requests. This is going to be tracked in #1075.